fix(signal): keep config compat under lint cap
